INFO STORE: clarify memory fields and fix int32 overflow
The INFO STORE store-stats had two problems: Log.MemorySizeBytes overflowed
int32 for in-memory logs larger than 2 GiB (BufferSize*PageSize and
AllocatedPageCount<<bits were evaluated in 32-bit before widening to long),
so e.g. a 128 GiB log reported a negative number; and several byte-valued
fields were named ambiguously (IndexMemorySize reported #cache-lines, not
bytes; BufferSize is a page count, not a byte size).

- Fix the int32 overflow in AllocatorBase.MaxMemorySizeBytes and
  LogAccessor.MemorySizeBytes/MemorySizeBytesIncludingOverflowPages (also
  fixes the INFO MEMORY store-memory aggregate that summed the negative).
- Report byte-valued fields in bytes with a Bytes suffix and split max vs
  current: IndexMemorySizeBytes; Log.PageSizeBytes, Log.MaxPageCount (was
  BufferSize), Log.MaxMemorySizeBytes, Log.CurrentMemorySizeBytes,
  Log.CurrentHeapSizeBytes (+ ReadCache mirror); grouped capacity then
  addresses. Added Tsavorite.IndexSizeBytes and LogAccessor.PageSizeBytes.
- Update the two affected tests (ConfigSetIndexSizeTest, GetStoreAddressInfo).

+++ libs/storage/Tsavorite/cs/src/core/Allocator/AllocatorBase.cs
@@ -811,7 +811,9 @@ namespace Tsavorite.core
         public int HighWaterAllocatedPageCount;
 
         /// <summary>Maximum memory size in bytes</summary>
-        public long MaxMemorySizeBytes => BufferSize * PageSize;
+        // (long) cast required: BufferSize and PageSize are both int, so the multiply would
+        // otherwise be evaluated in 32-bit and overflow (e.g. 2048 pages * 64 MiB = 128 GiB).
+        public long MaxMemorySizeBytes => (long)BufferSize * PageSize;
 
         /// <summary>Increments AllocatedPageCount. Updates MaxAllocatedPageCount if a higher number of pages have been allocated.</summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
